requirements.txt
PyJWT #==2.3.0
bcrypt #== 3.1.7
PyYAML #== 5.3.1
streamlit #== 1.37.0
extra-streamlit-components #== 0.1.70
streamlit_authenticator #== 0.4.1
